KU-118/refactor: /types/place.ts에 Place 관련 타입 구분
