+++ types/place.ts
@@ -10,6 +10,14 @@ export type KakaoPlace = {
   phone: string;
 };
 
+export type Place = {
+  placeName: string;
+  addressName: string;
+  roadAddressName: string;
+  latitude: number;
+  longitude: number;
+};
+
 export type MyLocation = {
   placeName: string;
   addressName: string;
